refactor(llm): extract provider error classification to _provider_errors
Moves the LLM error contract (LLMTruncationError, FatalLLMError,
_FATAL_HTTP_STATUS_CODES, _FATAL_MESSAGE_PATTERNS,
_LOGPROBS_UNSUPPORTED_PATTERNS, _TRANSIENT_LLM_EXCEPTION_NAMES,
_TRANSIENT_LLM_MESSAGE_TOKENS) plus the three classifier functions
(_classify_fatal_llm_error, _is_logprobs_unsupported_error,
_is_transient_llm_error) out of provider.py into a focused
amx/llm/_provider_errors.py.

provider.py re-exports the public names so callers
(amx.cli, analyze_flow, orchestrator, base agents, tests) keep
working byte-compatible. provider.py: 1733 -> 1504 LOC (-229, -13%).
